Big refactoring.  Note: this is part 1 of this checkin - I will need to a few.  Don't take this revision - wait for the one I say is safe to take (they should all be done by morning in Paris, so this probably won't matter).
Broke the project into two.  The remains of the old project, google-enterprise-connector-dctm, now builds without any dependency on documentum-licensed code.  To do this, I had to change some constants declared in interfaces.  The new project, dctm-client, contains the code that depends on documentum.  Unfortunately, most of the tests have to go in this project.  So we need more tests that don't depend on documentum!

The goal of this is to make it possible for us here to build the independent project and still run integration tests.  It should be possible for us to create a connector that uses only mocks.

Thanks all for your help with this.