Documentum Connector changes linked to Issue 32.
Analysis:

--------
When the timestamp field of the dm_audittrail table has the format
mm/dd/yyyy hh:mm:ss, in the query which retrieves documents to delete the
timestamp is formatted as yyyy-dd-mm ... instead of yyyy-mm-dd.

This problem could cause an infinite loop or make the connector never
return any document to delete.


Change Log:
----------

*  projects/dctm-client/source/java/com/google/enterprise/connector/dctm/dctmdfcwrap/DmTime.java
  - Added asString() method returns a String
  - Added time_pattern44 attribute represents the IDfTime.DF_TIME_PATTERN44 String pattern "mm/dd/yyyy hh:mi:ss" 
  - Added getTime_pattern44() getter returns the time_pattern44 attribute 
  
  
*  projects/dctm-client/source/java/com/google/enterprise/connector/dctm/dctmdfcwrap/DmCollection.java
  - Added getTime() method returns an ITime object