change log : ---------------------- * dctm-client\source\java\com\google\enterprise\connector\dctm\dctmdfcwrap\DmSysObject.java change to throw RepositoryDocumentException instead of RepositoryException
* dctm-client\source\java\com\google\enterprise\connector\dctm\dctmdfcwrap\DmQuery.java
execute method with arguments (Session, queryType) added

* dctm-client\source\java\com\google\enterprise\connector\dctm\dctmdfcwrap\DmSession.java
getObject method throws RepositoryDocumentException instead of RepositoryException

* dctm-client\source\java\com\google\enterprise\connector\dctm\dctmdfcwrap\DmSessionManager.java
sessionAdd, sessionDel and sessionAuto attributes added
setSessionAdd, setSessionDel, setSessionAuto, releasesessionAdd, releaseSessionDel and releaseSessionAuto added

* google-enterprise-connector-dctm\source\java\com\google\enterprise\connector\dctm\dfcwrap\IQuery.java
execute method with arguments (Session, queryType) added

* google-enterprise-connector-dctm\source\java\com\google\enterprise\connector\dctm\dfcwrap\ISessionManager.java
sessionAdd, sessionDel and sessionAuto attributes added
setSessionAdd, setSessionDel, setSessionAuto, releasesessionAdd, releaseSessionDel and releaseSessionAuto added

* google-enterprise-connector-dctm\source\java\com\google\enterprise\connector\dctm\dfcwrap\ISysObject.java
change to throw RepositoryDocumentException instead of RepositoryException

* google-enterprise-connector-dctm\source\java\com\google\enterprise\connector\dctm\DctmDocumentList.java
change in the checkpoint method to handle the case in which the fetch of the document could not be done.
change in the finalize method to handle the compatibility problem between DCTM 5.3 and 6

* google-enterprise-connector-dctm\source\java\com\google\enterprise\connector\dctm\DctmSysObjectDocument.java
throws RepositoryDocumentException instead of RepositoryException

* google-enterprise-connector-dctm\source\java\com\google\enterprise\connector\dctm\DctmAuthorizationManager.java
change in the authorizeDocids method to throw RepositoryException and to call setSessionAuto and releaseSessionAuto methods.

* google-enterprise-connector-dctm\source\java\com\google\enterprise\connector\dctm\DctmTraversalmanager.java
change in executeQuery method to call setSessionAdd and setSessionDel to handle the compatibility problem between DCTM 5.3 and 6.